refactor: setting modal refactored
