+++ src/components/cards/MyProfileCard.tsx
@@ -2,7 +2,7 @@ import { User } from '@/types/User';
 import NewGroupModal from '@/components/modals/group/new/NewGroupModal';
 import MySettingModal from '@/components/modals/my-setting/MySettingModal';
 import CustomLocationModal from '@/components/modals/my-setting/custom-location/CustomLocationModal';
-import ProfilePic from '@/components/ProfilePic';
+import ProfilePic from '@/components/cards/ProfilePic';
 import { SetCheckBtn } from '@/components/buttons';
 
 export default function MyProfileCard({ user }: { user: User }) {
