+++ src/components/modals/user-setting/contents/index.ts
@@ -1,0 +1,2 @@
+export * from './UserAddContent';
+export * from './UserDeleteContent';
